Allocate tuple with task to provide media for input data of tasks; Basic infrastructure for fence termination (in shared memory)

+++ parsec/ttg.cc
@@ -11,7 +11,7 @@
 using namespace parsec::ttg;
 using namespace ::ttg;
 
-using keyT = double;
+using keyT = uint64_t;
 
 class A : public Op<keyT, std::tuple<Out<keyT, int>, Out<keyT, int>>,
                        A, int> {
